\name{addCentromereBreaks}
\alias{addCentromereBreaks}
\title{
	Inserts breakpoints for a centromere
}
\description{
	Inserts breakpoints for centromeres in \code{RangedData} objects 
}
\usage{
addCentromereBreaks(ranges.object)
}
\arguments{
  \item{ranges.object}{
  An object of class code{RangedData}
}
}
\value{
	A \code{RangedData} object
}
\keyword{manip}
