\name{exampleData1}
\alias{exampleData1}
\docType{data}
\title{
	An object of class CNSet containing 16 randomly selected SNPS
}
\description{
	An object of class CNSet containing 16 randomly selected SNPS
}
\usage{data(exampleData1)}
\source{
	HapMap phase 3
}
\examples{
data(exampleData1)
show(exampleData1)
}
\keyword{datasets}
