\name{genotypeSet}
\alias{genotypeSet}
\docType{data}
\title{
	An object of class CNSet for 16 randomly selected SNPs in the
	HapMap Phase 3 dataset.
}
\description{
An object of cl	ass CNSet for 16 randomly selected SNPs in the
   	HapMap Phase 3 dataset.
}
\usage{data(genotypeSet)}

\examples{
data(genotypeSet)
## maybe str(genotypeSet) ; plot(genotypeSet) ...
}
\keyword{datasets}
