\name{redonSet}
\alias{redonSet}
\docType{data}
\title{
	Copy number estimates and genotype calls for chromosome 8 of HapMap sample NA19007.
}
\description{
	The object is of class \code{oligoSnpSet}
}
\usage{data(redonSet)}
\references{
 The amplification on chromosome 8p was original described by Redon et
 al., 2006
}
\examples{
data(redonSet)
## maybe str(redonSet) ; plot(redonSet) ...
}
\keyword{datasets}
